exchange/exchange/serializers.py
from rest_framework import serializers

from exchange.models import Provider


class ProviderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Provider
        fields = '__all__'
