requirements.txt
Django~=3.0
requests~=2.22.0
djangorestframework~=3.9.2
django-redis~=4.11.0
pyyaml~=5.1